Move core to own file
